Import version previously only maintained in FreeBSD
This repository starts with the version in /usr/ports/Tools/scripts of
the FreeBSD ports system committed on Sun May 22 22:13:41 2022 +0200
as commit 1f357cdb925.

This is not the initial version (which had been committed to the tools
directory in FreeBSD-14-CURRENT). This repository is only meant to be
used for further development of this tool before it is committed to
the FreeBSD ports tree.